Add a CLI client for the wifiplug protocol
Way too dump CLI client for the wifiplug protocol that allows talking to
the server using a username, password and then parse status messages and
send idle commands. It is also capable of parsing a JSON file to obtain
commands to send to the server for switching on/off plug allowing an
elaborate scheduling system to be set in place. No code to generate the
JSON is provided, feel free to create your own. Finally the output of
status messages is saved to a CSV file
Note: demo account is hardcoded (yey!!!)