Fix bug with stacktrace when plug is not online
When a plug is not online, the server will not return a status message.
Blocking on waiting for it will cause a contention between waiting to
receive it and the idle command, finally causing a stacktrace. Fix this
by using select(2) on status. Also cleanup some messages and make them
more informative